requirements.txt
langchain==0.3.19
langchain-anthropic==0.3.5
langchain-community==0.3.18
langchain-core==0.3.43
langchain-groq==0.2.4
langchain-openai==0.3.3
langchain-text-splitters==0.3.6
langdetect==1.0.9
langfuse==2.44.0
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.4
litellm==1.61.20
llama-cloud==0.1.9
llama-index==0.12.11
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.11
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.13
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
llvmlite==0.44.0
mcp==1.5.0
mistralai==1.6.0
ml-dtypes==0.3.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nano-graphrag==0.0.8.2
nano-vectordb==0.0.4.3
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
ollama==0.4.4
onnxruntime==1.21.0
open-webui==0.5.20
openai==1.61.0
openpyxl==3.1.5
opensearch-py==2.8.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==23.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pdf2image==1.17.0
pgvector==0.3.5
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pylint==3.3.4
pytest==8.3.5
pytest-docker==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.2
python-pptx==1.0.0
python-socketio==5.11.3
pytrec-eval-terrier==0.5.6
pytube==15.0.0
pytz==2024.2
pyventus==0.6.0
pyxlsb==1.0.10
PyYAML==6.0.2
qwen-vl-utils==0.0.10
rapidocr-onnxruntime==1.3.24
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.7
RestrictedPython==8.0
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
tzdata==2024.2
tzlocal==5.3.1
umap-learn==0.5.7
unstructured==0.16.17
unstructured-client==0.31.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
xlsx2csv==0.8.4
XlsxWriter==3.2.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
zstandard==0.23.0
